fix bug casting to_f instead of to_i

+++ lib/dm-pg-types/integer_array.rb
@@ -8,7 +8,7 @@ module DataMapper
       end
 
       def load(value)
-        super.map(&:to_f) if value
+        super.map(&:to_i) if value
       end
     end # class IntegerArray
   end # class Property
